Increase time limit for modal tests (#947)
## Summary
<!--- This is a required section; please describe the main purpose of
this proposed code change. --->
Modal tests on nvidia gpus failing becuase of time limit errors:
https://github.com/linkedin/Liger-Kernel/actions/runs/19472103084/job/55721981739.

This PR increases the timeout to 90 mins to fix this.

<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->

<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type: <BLANK>
- [ ] run `make test` to ensure correctness
- [ ] run `make checkstyle` to ensure code style
- [ ] run `make test-convergence` to ensure convergence