Olmo3 model support [ready for review] (#946)
## Summary
Add support for Olmo 3 models:
https://huggingface.co/allenai/Olmo-3-1125-32B

## Details
Olmo 3 is similar to Olmo 3 in all of the ways that matter to
Liger-kernel. It is a separate model type in `transformers` (in order to
support SWA) so instead of just aliasing Olmo 3 to Olmo 2 I add a new
model type here.

## Testing Done
Hardware Type: NVIDIA RTX 4090
- [x] run `make test` to ensure correctness
- [x] run `make checkstyle` to ensure code style
- [x] run `make test-convergence` to ensure convergence

---------

Co-authored-by: Vaibhav Jindal <vaibhav.jndl@gmail.com>